--- src/test/java/FullReadAndWriteTest.java
@@ -1,26 +1,0 @@
-import org.junit.jupiter.api.Test;
-
-public class FullReadAndWriteTest {
-
-    @Test
-    void 전체파일읽기_통과(){
-        
-    }
-
-    @Test
-    void 전체파일읽기_실패(){
-
-    }
-    
-    @Test
-    void 전체파일쓰기_10자리_입력_통과(){
-        
-    }
-
-
-    @Test
-    void 전체파일쓰기_10자리아님_입력_실패(){
-        
-    }
-
-}
