+++ src/test/java/SSDTest.java
@@ -5,13 +5,6 @@ import static org.junit.jupiter.api.Assertions.*;
 class SSDTest {
 
     @Test
-    void shell_exit_함수호출시_shell_종료확인() {
-    }
-
-    @Test
-    void shell_help_함수호출시_사용방법_출력확인() {
-    }
-
     void write_SSD_데이터_없는_곳에_쓰기_성공() {
     }
