+++ src/test/java/ShellTest.java
@@ -42,7 +42,31 @@ class ShellTest {
         //• 어떠한 명령어를 입력하더라도 Runtime Error가 나오면 안된다.
     }
 
+    @Test
+    void shell_exit_함수호출시_shell_종료확인() {
+    }
 
+    @Test
+    void shell_help_함수호출시_사용방법_출력확인() {
+    }
 
+    @Test
+    void 전체파일읽기_통과(){
 
+    }
+
+    @Test
+    void 전체파일읽기_실패(){
+
+    }
+
+    @Test
+    void 전체파일쓰기_10자리_입력_통과(){
+
+    }
+
+    @Test
+    void 전체파일쓰기_10자리아님_입력_실패(){
+
+    }
 }
